JARVIS IA/jarvis_backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
requests
numpy
scikit-learn
sentence-transformers
pandas
openpyxl
python-docx
PyMuPDF
fpdf2
joblib
passlib[bcrypt]
python-jose[cryptography]
supabase
email-validator